refactor: Put GeneralSettings Components under GeneralTab/Components folder
