+++ src/pages/Settings/components/GeneralTab/components/SettingsSwitch/index.tsx
